requirements.txt
playwright
beautifulsoup4
python-dotenv
chromadb
sentence-transformers
openai